+++ model/bandwidth-part-ue.h
@@ -74,8 +74,8 @@ public:
    */
   void SetMac (Ptr<MmWaveUeMac> s);
 
-  virtual void SetDlBandwidth (uint8_t bw) override { m_dlBandwidth = bw; }
-  virtual void SetUlBandwidth (uint8_t bw) override { m_ulBandwidth = bw; }
+  virtual void SetDlBandwidth (uint16_t bw) override { m_dlBandwidth = bw; }
+  virtual void SetUlBandwidth (uint16_t bw) override { m_ulBandwidth = bw; }
 
 protected:
   // inherited from Object
